+++ earth2studio/models/px/__init__.py
@@ -31,6 +31,7 @@ from earth2studio.models.px.aurora1p5 import (
     Aurora1p5Ensemble_6h,
 )
 from earth2studio.models.px.base import PrognosticModel
+from earth2studio.models.px.camulator import CAMulator
 from earth2studio.models.px.cbottle_video import CBottleVideo
 from earth2studio.models.px.datareplay import DataReplay
 from earth2studio.models.px.dlesym import DLESyM, DLESyMLatLon